Draft the design document and prepare a rough mockup of the C++ API (#1)
The most important part at the moment is the firmware update. You will
notice that my proposal differs from what is suggested in the work
order. I would like to discuss this in practical terms.

The rest seems mostly straightforward; I have attempted to capture the
main idea of that modem daemon API in my design.

---------

Co-authored-by: Sergei Shirokov <sergei.shirokov@zubax.com>